requirements.txt
Langchain
huggingface_hub
python-pptx
langchain-huggingface
langchain_community
transformers
streamlit

diffusers
accelerate
invisible_watermark
safetensors

langchainhub
wikipedia
langchain_ollama